chore(autoscaling): allow configuration of autoscaling policies
This commit adds the ability to configure autoscaling policies for the ECS
service. The module now accepts a list of objects, each representing a policy.

By default, the module creates a policy that scales based on the average CPU
utilization of the ECS service. The target value is set to 70% (old behavior).